borg: widen SPIR-B const_vals wire format from uint16_t to uint32_t
FP32-datapath plan item 6: software/borg/borg_spirb.h's const_vals[] was
uint16_t, blocking a real FP32 shader constant from ever reaching the GPR
file at full width. Widened the struct field, the byte-stream parser
(spirb_parse in borg_spirb.c, now reads 4 bytes per const instead of 2),
docs/spirb.md's format spec, and test_spirb.c's hand-built test blob to
match.

Wire-format-only for now, matching the companion mesa-side commit
(feat/fp32-datapath-borgc branch, not yet pushed): borgc has no FP32
codegen path, so every producer still writes an FP16 bit pattern
zero-extended into the low 16 bits. BORG_GPU->gpr[] (borg_driver.c) and
the RDL-generated gpr[32] array were already uint32_t, so no change
needed on the consuming side -- this was purely the wire format's own
width that was too narrow.

Verified: software/borg native test suite (test_spirb.c, 15/15).

Note: this repo's mesa submodule pointer is intentionally NOT bumped
here -- the companion borgc commit lives only on a local, unpushed
feat/fp32-datapath-borgc branch in this worktree's mesa checkout.
Bumping the pointer before that's pushed would break `git submodule
update` everywhere else (see this repo's own CLAUDE.md convention).

+++ software/borg/borg_spirb.h
@@ -23,7 +23,14 @@ typedef struct spirb_shader_t {
   uint8_t  attribute_regs[SPIRB_MAX_REGS];
   uint8_t  output_regs[SPIRB_MAX_REGS];
   uint8_t  const_regs[SPIRB_MAX_REGS];
-  uint16_t const_vals[SPIRB_MAX_REGS];
+  // uint32_t (widened from uint16_t): wire-format-only widening so the
+  // SPIR-B constant pool can eventually carry a real FP32 constant -- see
+  // mesa's encode.rs (emit_blob) doc comment, same commit. borgc has no
+  // FP32 codegen path yet, so every producer today still writes an FP16 bit
+  // pattern zero-extended into the low 16 bits; the firmware side (this
+  // parser, borg_kernel.c's GPR write) is format-agnostic either way, it
+  // just copies whatever width-correct value it's given into the GPR.
+  uint32_t const_vals[SPIRB_MAX_REGS];
 } spirb_shader_t;
 
 // Parse a SPIR-B blob from a byte array.
